feat: Refactored SelectSmart creating a componente to multiple header
